neues eingefügt

+++ app/Http/Controllers/ThreadsController.php
@@ -88,7 +88,18 @@ class ThreadsController extends Controller
 
     public function destroy($channel, Thread $thread){
 
+        $this->authorize('update', $thread);
+
+        if ($thread->user_id != auth()->id()){
+              abort(403, 'You do not have permission to do this');
+        }
         $thread->delete();
+
+        if(request()->wantsJson()){
+            return response([], 204);
+        }
+        return redirect('/threads');
+
 
     }
